@@ -32,46 +32,46 @@ namespace NBitcoin.Tests
 			Assert.Equal("129mVqKUmJ9uwPxKJBnNdABbuaaNfho4Ha", url.Address.ToString());
 			Assert.Equal(Money.Parse("0.06"), url.Amount);
 
-			url = new BitcoinUrlBuilder("bitcoin:129mVqKUmJ9uwPxKJBnNdABbuaaNfho4Ha?amount=0.06&label=Tom%20%26%20Jerry");
+			url = new BitcoinUrlBuilder("bitcoin:129mVqKUmJ9uwPxKJBnNdABbuaaNfho4Ha?amount=0.06&label=Tom%20%26%20Jerry", Network.Main);
 			Assert.Equal("129mVqKUmJ9uwPxKJBnNdABbuaaNfho4Ha", url.Address.ToString());
 			Assert.Equal(Money.Parse("0.06"), url.Amount);
 			Assert.Equal("Tom & Jerry", url.Label);
-			Assert.Equal(url.ToString(), new BitcoinUrlBuilder(url.ToString()).ToString());
+			Assert.Equal(url.ToString(), new BitcoinUrlBuilder(url.ToString(), Network.Main).ToString());
 
 			//Request 50 BTC with message: 
-			url = new BitcoinUrlBuilder("bitcoin:129mVqKUmJ9uwPxKJBnNdABbuaaNfho4Ha?amount=50&label=Luke-Jr&message=Donation%20for%20project%20xyz");
+			url = new BitcoinUrlBuilder("bitcoin:129mVqKUmJ9uwPxKJBnNdABbuaaNfho4Ha?amount=50&label=Luke-Jr&message=Donation%20for%20project%20xyz", Network.Main);
 			Assert.Equal(Money.Parse("50"), url.Amount);
 			Assert.Equal("Luke-Jr", url.Label);
 			Assert.Equal("Donation for project xyz", url.Message);
-			Assert.Equal(url.ToString(), new BitcoinUrlBuilder(url.ToString()).ToString());
+			Assert.Equal(url.ToString(), new BitcoinUrlBuilder(url.ToString(), Network.Main).ToString());
 
 			//Some future version that has variables which are (currently) not understood and required and thus invalid: 
-			url = new BitcoinUrlBuilder("bitcoin:129mVqKUmJ9uwPxKJBnNdABbuaaNfho4Ha?amount=50&label=Luke-Jr&message=Donation%20for%20project%20xyz&unknownparam=lol");
+			url = new BitcoinUrlBuilder("bitcoin:129mVqKUmJ9uwPxKJBnNdABbuaaNfho4Ha?amount=50&label=Luke-Jr&message=Donation%20for%20project%20xyz&unknownparam=lol", Network.Main);
 
 			//Some future version that has variables which are (currently) not understood but not required and thus valid: 
-			Assert.Throws<FormatException>(() => new BitcoinUrlBuilder("bitcoin:129mVqKUmJ9uwPxKJBnNdABbuaaNfho4Ha?amount=50&label=Luke-Jr&message=Donation%20for%20project%20xyz&req-unknownparam=lol"));
-			Assert.Throws<FormatException>(() => new BitcoinUrlBuilder("bitcoin:129mVqKUmJ9uwPxKJBnNdABbuaaNfho4Ha?amount=50&amount=50"));
+			Assert.Throws<FormatException>(() => new BitcoinUrlBuilder("bitcoin:129mVqKUmJ9uwPxKJBnNdABbuaaNfho4Ha?amount=50&label=Luke-Jr&message=Donation%20for%20project%20xyz&req-unknownparam=lol", Network.Main));
+			Assert.Throws<FormatException>(() => new BitcoinUrlBuilder("bitcoin:129mVqKUmJ9uwPxKJBnNdABbuaaNfho4Ha?amount=50&amount=50", Network.Main));
 
-			url = new BitcoinUrlBuilder("bitcoin:mq7se9wy2egettFxPbmn99cK8v5AFq55Lx?amount=0.11&r=https://merchant.com/pay.php?h%3D2a8628fc2fbe");
+			url = new BitcoinUrlBuilder("bitcoin:mq7se9wy2egettFxPbmn99cK8v5AFq55Lx?amount=0.11&r=https://merchant.com/pay.php?h%3D2a8628fc2fbe", Network.TestNet);
 			Assert.Equal("bitcoin:mq7se9wy2egettFxPbmn99cK8v5AFq55Lx?amount=0.11&r=https://merchant.com/pay.php?h%3d2a8628fc2fbe", url.ToString());
 #pragma warning disable CS0618 // Type or member is obsolete
 			Assert.Equal("https://merchant.com/pay.php?h=2a8628fc2fbe", url.PaymentRequestUrl.ToString());
 #pragma warning restore CS0618 // Type or member is obsolete
-			Assert.Equal(url.ToString(), new BitcoinUrlBuilder(url.ToString()).ToString());
+			Assert.Equal(url.ToString(), new BitcoinUrlBuilder(url.ToString(), Network.TestNet).ToString());
 
 			//Support no address
-			url = new BitcoinUrlBuilder("bitcoin:?r=https://merchant.com/pay.php?h%3D2a8628fc2fbe");
+			url = new BitcoinUrlBuilder("bitcoin:?r=https://merchant.com/pay.php?h%3D2a8628fc2fbe", Network.Main);
 #pragma warning disable CS0618 // Type or member is obsolete
 			Assert.Equal("https://merchant.com/pay.php?h=2a8628fc2fbe", url.PaymentRequestUrl.ToString());
 #pragma warning restore CS0618 // Type or member is obsolete
-			Assert.Equal(url.ToString(), new BitcoinUrlBuilder(url.ToString()).ToString());
+			Assert.Equal(url.ToString(), new BitcoinUrlBuilder(url.ToString(), Network.Main).ToString());
 		}
 
 		[Fact]
 		[Trait("UnitTest", "UnitTest")]
 		public void BitcoinUrlKeepUnknowParameter()
 		{
-			BitcoinUrlBuilder url = new BitcoinUrlBuilder("bitcoin:?r=https://merchant.com/pay.php?h%3D2a8628fc2fbe&idontknow=test");
+			BitcoinUrlBuilder url = new BitcoinUrlBuilder("bitcoin:?r=https://merchant.com/pay.php?h%3D2a8628fc2fbe&idontknow=test", Network.Main);
 
 			Assert.Equal("test", url.UnknowParameters["idontknow"]);
 			Assert.Single(url.UnknowParameters);
@@ -79,9 +79,9 @@ namespace NBitcoin.Tests
 
 		private BitcoinUrlBuilder CreateBuilder(string uri)
 		{
-			var builder = new BitcoinUrlBuilder(uri);
+			var builder = new BitcoinUrlBuilder(uri, Network.Main);
 			Assert.Equal(builder.Uri.ToString(), uri);
-			builder = new BitcoinUrlBuilder(new Uri(uri, UriKind.Absolute));
+			builder = new BitcoinUrlBuilder(new Uri(uri, UriKind.Absolute), Network.Main);
 			Assert.Equal(builder.ToString(), uri);
 			return builder;
 		}
