@@ -20,26 +20,28 @@ namespace NBitcoin.JsonConverters
 		{
 			settings.Converters.Add(new MoneyJsonConverter());
 			settings.Converters.Add(new KeyJsonConverter());
-			settings.Converters.Add(new CoinJsonConverter(network));
+			if (network != null)
+				settings.Converters.Add(new CoinJsonConverter(network));
 			settings.Converters.Add(new ScriptJsonConverter());
 			settings.Converters.Add(new FeeRateJsonConverter());
 			settings.Converters.Add(new UInt160JsonConverter());
 			settings.Converters.Add(new UInt256JsonConverter());
 			settings.Converters.Add(new LockTimeJsonConverter());
-			settings.Converters.Add(new PSBTJsonConverter(network));
+			settings.Converters.Add(new SequenceJsonConverter());
+			if (network != null)
+				settings.Converters.Add(new PSBTJsonConverter(network));
 			settings.Converters.Add(new HDFingerprintJsonConverter());
 			settings.Converters.Add(new OutpointJsonConverter());
-			settings.Converters.Add(new BitcoinSerializableJsonConverter(network));
+			if (network != null)
+				settings.Converters.Add(new BitcoinSerializableJsonConverter(network));
 			settings.Converters.Add(new NetworkJsonConverter());
 			settings.Converters.Add(new KeyPathJsonConverter());
 			settings.Converters.Add(new SignatureJsonConverter());
 			settings.Converters.Add(new HexJsonConverter());
 			settings.Converters.Add(new DateTimeToUnixTimeConverter());
 			settings.Converters.Add(new TxDestinationJsonConverter());
-			settings.Converters.Add(new BitcoinStringJsonConverter()
-			{
-				Network = network
-			});
+			if (network != null)
+				settings.Converters.Add(new BitcoinStringJsonConverter(network));
 			settings.ContractResolver = new CamelCasePropertyNamesContractResolver();
 		}
 
