@@ -1,5 +1,4 @@
-﻿using NBitcoin.BitcoinCore;
-using NBitcoin.Policy;
+﻿using NBitcoin.Policy;
 using System;
 using System.Linq;
 
@@ -44,34 +43,6 @@ namespace NBitcoin
 				return false;
 
 			return template.CheckScriptSig(scriptSig, scriptPubKey);
-		}
-
-		//
-		// Check transaction inputs, and make sure any
-		// pay-to-script-hash transactions are evaluating IsStandard scripts
-		//
-		// Why bother? To avoid denial-of-service attacks; an attacker
-		// can submit a standard HASH... OP_EQUAL transaction,
-		// which will get accepted into blocks. The redemption
-		// script can be anything; an attacker could use a very
-		// expensive-to-check-upon-redemption script like:
-		//   DUP CHECKSIG DROP ... repeated 100 times... OP_1
-		[Obsolete]
-		public static bool AreInputsStandard(Transaction tx, CoinsView coinsView)
-		{
-			if (tx.IsCoinBase)
-				return true; // Coinbases don't use vin normally
-
-			foreach (var input in tx.Inputs)
-			{
-				TxOut prev = coinsView.GetOutputFor(input);
-				if (prev == null)
-					return false;
-				if (!IsStandardScriptSig(input.ScriptSig, prev.ScriptPubKey))
-					return false;
-			}
-
-			return true;
 		}
 	}
 }
